# openapi/api-docs/openapi3_0 (1).yaml
openapi: 3.0.1
info:
  title: "Project Elara: e-commerce platform"
  description: "Initial version of public API. It may change while development process. Contains final version of Login, Users and Products GET services."
  version: v1
servers:
  - url: http://localhost:8080/api
    description: Gateway root url
paths:
  /v1/auth/user:
    get:
      tags:
        - login-controller
      summary: Get user info by token.
      description: Get user info by token.
      operationId: getUserByToken
      parameters:
        - name: token
          in: query
          description: JWT token
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          description: service token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Return user info
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserServiceInfoDTO'
        '400':
          description: Token is not provided
          content:
            '*/*':
              schema:
                type: string
        '401':
          description: Invalid service token
          content:
            '*/*':
              schema:
                type: string
        '500':
          description: Login service is unavailable
          content:
            '*/*':
              schema:
                type: string
  /v1/auth/login:
    get:
      tags:
        - login-controller
      summary: Perform login operation.
      description: Perform login operation.
      operationId: login
      parameters:
        - name: login
          in: query
          description: user login (should match email validation rules)
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: user password
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Login successful
          content:
            '*/*':
              schema:
                type: object
        '400':
          description: Login or password is not provided
          content:
            '*/*':
              schema:
                type: string
        '401':
          description: Login or password is incorrect
          content:
            '*/*':
              schema:
                type: string
        '500':
          description: Login service is unavailable
          content:
            '*/*':
              schema:
                type: string
  /v1/users:
    put:
      tags:
        - users-protected-controller
      summary: Update user profile
      description: |-
        Update user profile.
         <p>
         Note: This method is protected by JWT token.
      operationId: updateProfile
      parameters:
        - name: Authorization
          in: header
          description: JWT token
          required: true
          schema:
            type: string
      requestBody:
        description: user profile to update
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdateRequestDTO'
        required: true
      responses:
        '204':
          description: Profile updated
        '400':
          description: Invalid profile
          content:
            '*/*':
              schema:
                type: string
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                type: string
    post:
      tags:
        - users-public-controller
      summary: Register new user
      description: Register new user
      operationId: register
      requestBody:
        description: user info
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequestDTO'
        required: true
      responses:
        '201':
          description: User registered
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserInfoDTO'
        '400':
          description: Invalid user info
          content:
            '*/*':
              schema:
                type: string
        '409':
          description: Login is unavailable
          content:
            '*/*':
              schema:
                type: string
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                type: string
  /v1/users/change-password:
    put:
      tags:
        - users-protected-controller
      summary: Change password request
      description: |-
        Change password request.
         <p>
         Note: This method is protected by JWT token.
      operationId: changePassword
      parameters:
        - name: Authorization
          in: header
          description: JWT token
          required: true
          schema:
            type: string
      requestBody:
        description: request with old and new password
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequestDTO'
        required: true
      responses:
        '204':
          description: Password changed
        '400':
          description: Invalid password
          content:
            '*/*':
              schema:
                type: string
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                type: string
...
  /v1/users/{userId}:
    get:
      tags:
        - users-protected-controller
      summary: Get user info by id
      description: |-
        Get user info by id.
         <p>
         Note: This method is service (protected by service token).
      operationId: getUserInfoById
      parameters:
        - name: userId
          in: path
          description: user id
          required: true
          schema:
            type: string
            format: uuid
        - name: Authorization
          in: header
          description: service token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User info found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserInfoDTO'
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
        '404':
          description: User info not found
          content:
            '*/*':
              schema:
                type: string
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                type: string
    delete:
      tags:
        - users-protected-controller
      summary: Delete user by id
      description: |-
        Delete user by id.
         <p>
         Note: This method is protected by JWT token.
      operationId: deleteUserById
      parameters:
        - name: userId
          in: path
          description: user id
          required: true
          schema:
            type: string
            format: uuid
        - name: Authorization
          in: header
          description: JWT token
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User deleted
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
        '404':
          description: User not found
          content:
            '*/*':
              schema:
                type: string
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                type: string
  /v1/users/profile/{userId}:
    get:
      tags:
        - users-protected-controller
      summary: Get user profile by id
      description: |-
        Get user profile by id.
         <p>
         Note: This method is protected by JWT token.
      operationId: getUserProfileById
      parameters:
        - name: userId
          in: path
          description: user id
          required: true
          schema:
            type: string
            format: uuid
        - name: Authorization
          in: header
          description: JWT token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Profile found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserProfileDTO'
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
        '404':
          description: Profile not found
          content:
            '*/*':
              schema:
                type: string
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                type: string
...
components:
  schemas:
    UserServiceInfoDTO:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        login:
          type: string
        roleId:
          type: string
          format: uuid
        roleName:
          type: string
          enum:
            - EMAIL_NOT_VERIFIED
            - USER
            - ADMIN
      description: ''
    UserProfileUpdateRequestDTO:
      required:
        - email
        - firstName
        - lastName
        - userId
      type: object
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        pictureUrl:
          type: string
        birthDate:
          type: string
          format: date-time
      description: ''
    ChangePasswordRequestDTO:
      required:
        - newPassword
        - oldPassword
        - userId
      type: object
      properties:
        userId:
          type: string
          format: uuid
        oldPassword:
          type: string
        newPassword:
          type: string
      description: ''
    UserRegisterRequestDTO:
      required:
        - firstName
        - lastName
        - login
        - password
      type: object
      properties:
        login:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        pictureUrl:
          type: string
        birthDate:
          type: string
          format: date-time
      description: ''
    UserInfoDTO:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
        firstName:
          type: string
      description: ''
...
    UserProfileDTO:
      type: object
      properties:
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        pictureUrl:
          type: string
        birthDate:
          type: string
          format: date-time
      description: ''
...